fix: have timeMachine flux editor use the common position calculations
